socket still no good

+++ myftp/stream.h
@@ -53,7 +53,7 @@ int read_code(int fd,char *code);
  */
 int write_twobytelength(int fd, int length);
 
-int read_twobytelength(int fd, short* length);
+int read_twobytelength(int fd, int* length);
 
 
 /* Writes a four byte integer from length to socket fd.
